+++ src/modules/api/external-api.service.ts
@@ -3,6 +3,8 @@ import { SpoonacularApiService } from './spoonacular/spoonacular.api.service';
 import { AbstractApiService } from '../../services/abstract.api.service';
 import { DetailedDish, RatedDish } from '../dish/dish.types';
 import { MealType } from '../../common/enums';
+import { Language } from '../../common/types';
+import { DishRecipe } from '../recipe/recipe.types';
 
 @Injectable()
 export class ExternalApiService {
@@ -22,5 +24,9 @@ export class ExternalApiService {
 
     getDishDetails(id: string): Promise<DetailedDish>[] {
         return this.getAll().map(service => service.getDishDetails(id));
+    }
+
+    getDishRecipe(dishId: string, language: Language): Promise<DishRecipe | null>[] {
+        return this.getAll().map(service => service.getDishRecipe(dishId, language));
     }
 }