+++ src/modules/recipe/recipe.types.ts
@@ -1,0 +1,8 @@
+import { Language } from '../../common/types';
+import { DishRecipeSections } from '../dish/dish.types';
+
+export interface DishRecipe {
+    language: Language;
+    dishId: string;
+    sections: DishRecipeSections;
+}