+++ src/modules/translation/translation.service.ts
@@ -1,34 +1,28 @@
 import { Injectable } from '@nestjs/common';
 import { Language } from '../../common/types';
-import { TranslatedDetailedDish, TranslatedIngredient } from './translation.types';
-import { DetailedDish, DishRecipeSections } from '../dish/dish.types';
+import { TranslatedDetailedDish, TranslatedIngredient, TranslatedRecipe } from './translation.types';
+import { DetailedDish } from '../dish/dish.types';
 import { compoundTextToTranslate, convertAmountToText, normalizeName, normalizeUnit } from '../../common/helpers';
 import translate from '@iamtraction/google-translate';
 import { proceedTagsSpaces } from '../api/spoonacular/spoonacular.api.utils';
 import { DishIngredient } from '../ingredient/ingredient.types';
+import { DishRecipe } from '../recipe/recipe.types';
 
 @Injectable()
 export class TranslationService {
 
-    async translateDish(dish: DetailedDish, targetLanguage: Language): Promise<TranslatedDetailedDish> {
-        if (dish.language === targetLanguage) {
-            const translatedIngredients = await this.translateIngredients(dish.ingredients, targetLanguage);
+    async translateDish(dish: DetailedDish, targetLanguage: Language): Promise<TranslatedDetailedDish | null> {
+        const { description, ingredients, language } = dish;
+
+        if (language === targetLanguage) {
+            const translatedIngredients = await this.translateIngredients(ingredients, { originalLanguage: dish.language, targetLanguage });
 
             return {
-                description: '',
-                ingredients: translatedIngredients,
-                recipe: []
+                description,
+                ingredients: translatedIngredients
             };
         }
 
-        const { description, ingredients, recipeSections } = dish;
-        const recipeNewSectionIndexes: number[] = recipeSections.reduce((acc, curr, idx) => {
-            if (acc.length === 0) {
-                return [0];
-            }
-
-            return [...acc, acc.at(-1) + recipeSections[idx - 1].steps.length + 1];
-        }, []);
         const ingredientImages: string[] = [];
         const startRecipeIdx: number = ingredients.length;
 
@@ -44,47 +38,26 @@ export class TranslationService {
                 ingredientImages.push(imageUrl);
 
                 return compoundedText;
-            }),
-            ...recipeSections.map(section => {
-                return [
-                    section.name,
-                    ...section.steps
-                ];
-            }).flat()
+            })
         ].join('\n');
 
         const translatedResult = await this.translate(stringToTranslate, targetLanguage);
         const [ translatedDescription, ...translatedIngredientsAndRecipes] = translatedResult.split('\n');
         const translatedIngredients = translatedIngredientsAndRecipes.slice(0, startRecipeIdx);
-        const translatedRecipes = translatedIngredientsAndRecipes.slice(startRecipeIdx);
 
         const ingredientList: TranslatedIngredient[] = translatedIngredients.map((ingredient, idx) => ({ text: ingredient, imageUrl: ingredientImages[idx] }));
-        const translatedRecipeSections: DishRecipeSections = recipeNewSectionIndexes.map((index, idx, sections) => {
-            const nextSection = sections[idx + 1];
-
-            if (!nextSection) {
-                return {
-                    name: translatedRecipes[index],
-                    steps: translatedRecipes.slice(index + 1)
-                };
-            }
-
-            return {
-                name: translatedRecipes[index],
-                steps: translatedRecipes.slice(index + 1, nextSection)
-            };
-        });
 
         return {
             description: proceedTagsSpaces(translatedDescription),
-            ingredients: ingredientList,
-            recipe: translatedRecipeSections
+            ingredients: ingredientList
         };
     }
 
-    async translateIngredients(ingredients: DishIngredient[], targetLanguage: Language): Promise<TranslatedIngredient[]> {
+    async translateIngredients(ingredients: DishIngredient[], options: { originalLanguage: Language, targetLanguage: Language }): Promise<TranslatedIngredient[]> {
+        const { originalLanguage, targetLanguage } = options;
+
         const ingredientImages: string[] = [];
-        const ingredientsToTranslate = ingredients.map(ingredient => {
+        const ingredientsList = ingredients.map(ingredient => {
             const { amount, unit, name, imageUrl } = ingredient;
             const normalizedName = normalizeName(name);
             const normalizedUnit = normalizeUnit(amount, unit);
@@ -94,8 +67,16 @@ export class TranslationService {
             ingredientImages.push(imageUrl);
 
             return compoundedText;
-        }).join('\n');
+        });
 
+        if (originalLanguage === targetLanguage) {
+            return ingredientsList.map((ingredientText, idx) => ({
+                text: ingredientText,
+                imageUrl: ingredientImages[idx]
+            }));
+        }
+
+        const ingredientsToTranslate = ingredientsList.join('\n');
         const translatedResult = await this.translate(ingredientsToTranslate, targetLanguage);
         const translatedIngredients = translatedResult.split('\n');
 
@@ -103,6 +84,43 @@ export class TranslationService {
             text: ingredientText,
             imageUrl: ingredientImages[idx]
         }));
+    }
+
+    async translateRecipe(recipe: DishRecipe, targetLanguage: Language): Promise<TranslatedRecipe> {
+        if (recipe.language === targetLanguage) {
+            return {
+                original: recipe,
+                translated: recipe
+            };
+        }
+
+        const sectionSeparator = '\t\t\t';
+        const innerSectionSeparator = '\n\n';
+        const stepSeparator = '\\';
+
+        const stringToTranslate = recipe.sections
+            .map(s => (s.steps.join(stepSeparator) + innerSectionSeparator + s.name))
+            .join(sectionSeparator);
+        const translatedResult = await this.translate(stringToTranslate, targetLanguage);
+        const sections = translatedResult.split(sectionSeparator);
+        const translatedRecipe = sections.map(s => {
+            const [stepString, name] = s.split(innerSectionSeparator);
+            const steps = stepString.split(stepSeparator);
+
+            return {
+                name: name ?? '',
+                steps
+            };
+        });
+
+        return {
+            original: recipe,
+            translated: {
+                language: targetLanguage,
+                dishId: recipe.dishId,
+                sections: translatedRecipe
+            }
+        };
     }
 
     async translate(text: string, targetLanguage: Language): Promise<string> {