+++ src/modules/user/user.types.ts
@@ -10,8 +10,6 @@ export type UserPermissions = Pick<UserDocument, 'isAdmin' | 'capabilities'>;
 
 export type UserObject = Pick<UserDocument, 'email' | 'login' | 'isAdmin' | 'capabilities'> & { id: string };
 
-export type AuthenticatedUserRequestBody<T> = { authenticatedUser: UserAccessTokenPayload, data: T };
-
 export type UserProfile = Pick<UserDocument, 'login' | 'isAdmin' | 'capabilities' | 'activated'> & {
     dishList: Array<Pick<DishDocument, 'title'> & { id: string }>;
 };