+++ src/mongodb/documents/dish.document.ts
@@ -1,6 +1,5 @@
 import { Document } from 'mongoose';
 import { DishIngredient } from '../../modules/ingredient/ingredient.types';
-import { DishRecipeSections } from '../../modules/dish/dish.types';
 import { Language } from '../../common/types';
 import { MealType, DishType } from '../../common/enums';
 
@@ -18,5 +17,4 @@ export interface DishDocument extends Document {
   readonly softEdited?: DishDocument;
   readonly softDeleted?: boolean;
   readonly readyInMinutes: number;
-  readonly recipeSections: DishRecipeSections;
 }