+++ src/services/test.api.service.ts
@@ -1,6 +1,6 @@
 import { AbstractApiService } from './abstract.api.service';
 import { ApiName } from '../modules/redis/redis.types';
-import { DetailedDish, RatedDish } from '../modules/dish/dish.types';
+import { DetailedDish, DishRecipeSections, RatedDish } from '../modules/dish/dish.types';
 import { IngredientType, DishIngredient } from '../modules/ingredient/ingredient.types';
 
 export class TestApiService extends AbstractApiService<any, any, any, any> {
@@ -41,8 +41,12 @@ export class TestApiService extends AbstractApiService<any, any, any, any> {
         return [...data];
     }
 
-    proceedDataToDishDetails(data: any, instructionData: any): DetailedDish {
-        return { ...data, ...instructionData };
+    proceedDataToDishDetails(data: any): DetailedDish {
+        return { ...data };
+    }
+
+    proceedDataToDishRecipeSections(instructionData: any): DishRecipeSections {
+        return { ...instructionData };
     }
 
 }